# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Rolls API",
    "version": "1.0.0",
    "description": "OpenAPI spec generated from Next.js route handlers."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local dev (API base)"
    },
    {
      "url": "https://rolls.yannick.computer/api",
      "description": "Production"
    }
  ],
  "paths": {
    "/auth/api-keys": {
      "get": {
        "operationId": "get-auth-api-keys",
        "summary": "List API keys",
        "description": "Lists API keys for the authenticated user (raw keys are never returned).",
        "tags": [
          "Auth"
        ],
        "parameters": [],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post-auth-api-keys",
        "summary": "Create API key",
        "description": "Creates a new API key for the authenticated user. `raw_key` is returned only once.",
        "tags": [
          "Auth"
        ],
        "parameters": [],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyCreateBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyCreateResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/webauthn/login-verify": {
      "post": {
        "operationId": "post-auth-webauthn-login-verify",
        "summary": "WebAuthn login verify",
        "description": "Verifies a WebAuthn authentication response and starts a session (via Set-Cookie).",
        "tags": [
          "Auth"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebAuthnLoginVerifyBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionAuthSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/webauthn/register-options": {
      "post": {
        "operationId": "post-auth-webauthn-register-options",
        "summary": "WebAuthn registration options",
        "description": "Returns WebAuthn registration options + the challenge to be echoed to verify.",
        "tags": [
          "Auth"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebAuthnRegisterOptionsBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebAuthnRegisterOptionsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/webauthn/register-verify": {
      "post": {
        "operationId": "post-auth-webauthn-register-verify",
        "summary": "WebAuthn registration verify",
        "description": "Verifies a WebAuthn registration response and starts a session (via Set-Cookie).",
        "tags": [
          "Auth"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebAuthnRegisterVerifyBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionAuthSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/cameras": {
      "get": {
        "operationId": "get-cameras",
        "summary": "List cameras",
        "description": "",
        "tags": [
          "Cameras"
        ],
        "parameters": [],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Camera"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post-cameras",
        "summary": "Create (or upsert) a camera",
        "description": "Slug is generated from brand+model. Upserts on (user_id, slug).",
        "tags": [
          "Cameras"
        ],
        "parameters": [],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCameraBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Camera"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiKeyListResponse": {
        "type": "object",
        "properties": {
          "api_keys": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "label": {
                  "type": "string",
                  "nullable": true
                },
                "last_used_at": {
                  "type": "string",
                  "nullable": true
                },
                "created_at": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "ApiKeyCreateBody": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ApiKeyCreateResponse": {
        "type": "object",
        "properties": {
          "api_key": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "label": {
                "type": "string",
                "nullable": true
              },
              "created_at": {
                "type": "string"
              }
            }
          },
          "raw_key": {
            "type": "string"
          }
        }
      },
...
      "WebAuthnLoginVerifyBody": {
        "type": "object",
        "properties": {
          "response": {
            "type": "object",
            "nullable": false
          },
          "challenge": {
            "type": "string",
            "nullable": false
          },
          "user_id": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "SessionAuthSuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "username": {
                "type": "string"
              },
              "name": {
                "type": "string",
                "nullable": true
              },
              "email": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "WebAuthnRegisterOptionsBody": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "nullable": false
          },
          "email": {
            "type": "string",
            "nullable": false
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "invite_code": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "WebAuthnRegisterOptionsResponse": {
        "type": "object",
        "properties": {
          "options": {
            "type": "object"
          },
          "challenge": {
            "type": "string"
          }
        }
      },
      "WebAuthnRegisterVerifyBody": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "nullable": false
          },
          "email": {
            "type": "string",
            "nullable": false
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "invite_code": {
            "type": "string",
            "nullable": true
          },
          "response": {
            "type": "object",
            "nullable": false
          },
          "challenge": {
            "type": "string",
            "nullable": false
          },
          "device_name": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "Camera": {},
      "CreateCameraBody": {},
...
    },
...
  }
}